style: forge fmt — fix all formatting to pass CI check
747 files reformatted. CI forge fmt --check will now pass.
1041 tests passing, 0 failures.

+++ contracts/fhe/utils/debug/MockFheOps.sol
@@ -48,10 +48,7 @@ contract MockFheOps {
         return result;
     }
 
-    function bytes32ToBytes(
-        bytes32 input,
-        uint8
-    ) internal pure returns (bytes memory) {
+    function bytes32ToBytes(bytes32 input, uint8) internal pure returns (bytes memory) {
         return bytes.concat(input);
     }
 
@@ -79,15 +76,9 @@ contract MockFheOps {
     }
 
     //for unknown size of bytes - we could instead just encode it as bytes32 because it's always uint256 but for now lets keep it like this
-    function bytesToUint(
-        bytes memory b
-    ) internal pure virtual returns (uint256) {
+    function bytesToUint(bytes memory b) internal pure virtual returns (uint256) {
         require(b.length <= 32, "Bytes length exceeds 32.");
-        return
-            abi.decode(
-                abi.encodePacked(new bytes(32 - b.length), b),
-                (uint256)
-            );
+        return abi.decode(abi.encodePacked(new bytes(32 - b.length), b), (uint256));
     }
 
     function bytesToBool(bytes memory b) internal pure virtual returns (bool) {
@@ -96,244 +87,139 @@ contract MockFheOps {
         return value != 0;
     }
 
-    function trivialEncrypt(
-        bytes memory input,
-        uint8 toType,
-        int32
-    ) external pure returns (bytes memory) {
+    function trivialEncrypt(bytes memory input, uint8 toType, int32) external pure returns (bytes memory) {
         // forge-lint: disable-next-line(unsafe-typecast)
         bytes32 result = bytes32(input);
         return bytes32ToBytes(result, toType);
     }
 
-    function add(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        uint256 result = (bytesToUint(lhsHash) + bytesToUint(rhsHash)) %
-            maxValue(utype);
+    function add(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        uint256 result = (bytesToUint(lhsHash) + bytesToUint(rhsHash)) % maxValue(utype);
         return uint256ToBytes(result);
     }
 
-    function sealOutput(
-        uint8,
-        bytes memory ctHash,
-        bytes memory
-    ) external pure returns (string memory) {
+    function sealOutput(uint8, bytes memory ctHash, bytes memory) external pure returns (string memory) {
         return string(ctHash);
     }
 
-    function verify(
-        uint8,
-        bytes memory input,
-        int32
-    ) external pure returns (bytes memory) {
+    function verify(uint8, bytes memory input, int32) external pure returns (bytes memory) {
         return input;
     }
 
-    function cast(
-        uint8,
-        bytes memory input,
-        uint8 toType
-    ) external pure returns (bytes memory) {
+    function cast(uint8, bytes memory input, uint8 toType) external pure returns (bytes memory) {
         // forge-lint: disable-next-line(unsafe-typecast)
         bytes32 result = bytes32(input);
         return bytes32ToBytes(result, toType);
     }
 
-    function log(string memory s) external pure {}
+    function log(string memory s) external pure { }
 
-    function decrypt(
-        uint8,
-        bytes memory input,
-        uint256
-    ) external pure returns (uint256) {
+    function decrypt(uint8, bytes memory input, uint256) external pure returns (uint256) {
         uint256 result = bytesToUint(input);
         return result;
     }
 
-    function lte(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        bool result = (bytesToUint(lhsHash) % maxValue(utype)) <=
-            (bytesToUint(rhsHash) % maxValue(utype));
+    function lte(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        bool result = (bytesToUint(lhsHash) % maxValue(utype)) <= (bytesToUint(rhsHash) % maxValue(utype));
         return boolToBytes(result);
     }
 
-    function sub(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        uint256 result = (bytesToUint(lhsHash) - bytesToUint(rhsHash)) %
-            maxValue(utype);
+    function sub(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        uint256 result = (bytesToUint(lhsHash) - bytesToUint(rhsHash)) % maxValue(utype);
         return uint256ToBytes(result);
     }
 
-    function mul(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        uint256 result = (bytesToUint(lhsHash) * bytesToUint(rhsHash)) %
-            maxValue(utype);
+    function mul(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        uint256 result = (bytesToUint(lhsHash) * bytesToUint(rhsHash)) % maxValue(utype);
         return uint256ToBytes(result);
     }
 
-    function lt(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        bool result = (bytesToUint(lhsHash) % maxValue(utype)) <
-            (bytesToUint(rhsHash) % maxValue(utype));
+    function lt(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        bool result = (bytesToUint(lhsHash) % maxValue(utype)) < (bytesToUint(rhsHash) % maxValue(utype));
         return boolToBytes(result);
     }
 
-    function select(
-        uint8,
-        bytes memory controlHash,
-        bytes memory ifTrueHash,
-        bytes memory ifFalseHash
-    ) external pure returns (bytes memory) {
+    function select(uint8, bytes memory controlHash, bytes memory ifTrueHash, bytes memory ifFalseHash)
+        external
+        pure
+        returns (bytes memory)
+    {
         bool control = bytesToBool(controlHash);
         if (control) return ifTrueHash;
         return ifFalseHash;
     }
 
-    function req(
-        uint8,
-        bytes memory input
-    ) external pure returns (bytes memory) {
+    function req(uint8, bytes memory input) external pure returns (bytes memory) {
         bool condition = (bytesToUint(input) != 0);
         require(condition);
         return input;
     }
 
-    function div(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        uint256 result = (bytesToUint(lhsHash) / bytesToUint(rhsHash)) %
-            maxValue(utype);
+    function div(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        uint256 result = (bytesToUint(lhsHash) / bytesToUint(rhsHash)) % maxValue(utype);
         return uint256ToBytes(result);
     }
 
-    function gt(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        bool result = (bytesToUint(lhsHash) % maxValue(utype)) >
-            (bytesToUint(rhsHash) % maxValue(utype));
+    function gt(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        bool result = (bytesToUint(lhsHash) % maxValue(utype)) > (bytesToUint(rhsHash) % maxValue(utype));
         return boolToBytes(result);
     }
 
-    function gte(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        bool result = (bytesToUint(lhsHash) % maxValue(utype)) >=
-            (bytesToUint(rhsHash) % maxValue(utype));
+    function gte(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        bool result = (bytesToUint(lhsHash) % maxValue(utype)) >= (bytesToUint(rhsHash) % maxValue(utype));
         return boolToBytes(result);
     }
 
-    function rem(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        uint256 result = (bytesToUint(lhsHash) % bytesToUint(rhsHash)) %
-            maxValue(utype);
+    function rem(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        uint256 result = (bytesToUint(lhsHash) % bytesToUint(rhsHash)) % maxValue(utype);
         return uint256ToBytes(result);
     }
 
-    function and(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
+    function and(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
         // forge-lint: disable-next-line(unsafe-typecast)
         bytes32 result = bytes32(lhsHash) & bytes32(rhsHash);
         return bytes32ToBytes(result, utype);
     }
 
-    function or(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
+    function or(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
         // forge-lint: disable-next-line(unsafe-typecast)
         bytes32 result = bytes32(lhsHash) | bytes32(rhsHash);
         return bytes32ToBytes(result, utype);
     }
 
-    function xor(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
+    function xor(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
         // forge-lint: disable-next-line(unsafe-typecast)
         bytes32 result = bytes32(lhsHash) ^ bytes32(rhsHash);
         return bytes32ToBytes(result, utype);
     }
 
-    function eq(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        bool result = (bytesToUint(lhsHash) % maxValue(utype)) ==
-            (bytesToUint(rhsHash) % maxValue(utype));
+    function eq(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        bool result = (bytesToUint(lhsHash) % maxValue(utype)) == (bytesToUint(rhsHash) % maxValue(utype));
         return boolToBytes(result);
     }
 
-    function ne(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        bool result = (bytesToUint(lhsHash) % maxValue(utype)) !=
-            (bytesToUint(rhsHash) % maxValue(utype));
+    function ne(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        bool result = (bytesToUint(lhsHash) % maxValue(utype)) != (bytesToUint(rhsHash) % maxValue(utype));
         return boolToBytes(result);
     }
 
-    function min(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        bool result = (bytesToUint(lhsHash) % maxValue(utype)) >=
-            (bytesToUint(rhsHash) % maxValue(utype));
+    function min(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        bool result = (bytesToUint(lhsHash) % maxValue(utype)) >= (bytesToUint(rhsHash) % maxValue(utype));
         if (result == true) {
             return rhsHash;
         }
         return lhsHash;
     }
 
-    function max(
-        uint8 utype,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
-        bool result = (bytesToUint(lhsHash) % maxValue(utype)) >=
-            (bytesToUint(rhsHash) % maxValue(utype));
+    function max(uint8 utype, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
+        bool result = (bytesToUint(lhsHash) % maxValue(utype)) >= (bytesToUint(rhsHash) % maxValue(utype));
         if (result == true) {
             return lhsHash;
         }
         return rhsHash;
     }
 
-    function shl(
-        uint8,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
+    function shl(uint8, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
         uint256 lhs = bytesToUint(lhsHash);
         uint256 rhs = bytesToUint(rhsHash);
 
@@ -342,11 +228,7 @@ contract MockFheOps {
         return uint256ToBytes(result);
     }
 
-    function shr(
-        uint8,
-        bytes memory lhsHash,
-        bytes memory rhsHash
-    ) external pure returns (bytes memory) {
+    function shr(uint8, bytes memory lhsHash, bytes memory rhsHash) external pure returns (bytes memory) {
         uint256 lhs = bytesToUint(lhsHash);
         uint256 rhs = bytesToUint(rhsHash);
 
@@ -355,29 +237,18 @@ contract MockFheOps {
         return uint256ToBytes(result);
     }
 
-    function not(
-        uint8 utype,
-        bytes memory value
-    ) external pure returns (bytes memory) {
+    function not(uint8 utype, bytes memory value) external pure returns (bytes memory) {
         // forge-lint: disable-next-line(unsafe-typecast)
         bytes32 result = ~bytes32(value);
         return bytes32ToBytes(result, utype);
     }
 
     function getNetworkPublicKey(int32) external pure returns (bytes memory) {
-        string
-            memory x = "((-(-_(-_-)_-)-)) You've stepped into the wrong neighborhood pal.";
+        string memory x = "((-(-_(-_-)_-)-)) You've stepped into the wrong neighborhood pal.";
         return bytes(x);
     }
 
-    function random(
-        uint8 utype,
-        uint64,
-        int32
-    ) external view returns (bytes memory) {
-        return
-            uint256ToBytes(
-                uint(keccak256(abi.encode(block.timestamp))) % maxValue(utype)
-            );
+    function random(uint8 utype, uint64, int32) external view returns (bytes memory) {
+        return uint256ToBytes(uint256(keccak256(abi.encode(block.timestamp))) % maxValue(utype));
     }
 }
